backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.0.32
python-dotenv==1.0.0
SQLAlchemy==2.0.4
bcrypt==4.0.1 